+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/angular
 
+## 2.0.248
+
+### Patch Changes
+
+- Updated dependencies [dab0a08]
+  - @ai-sdk/provider-utils@4.0.43
+  - ai@6.0.247
+
 ## 2.0.247
 
 ### Patch Changes