+++ packages/black-forest-labs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/black-forest-labs
 
+## 1.0.53
+
+### Patch Changes
+
+- Updated dependencies [dab0a08]
+  - @ai-sdk/provider-utils@4.0.43
+
 ## 1.0.52
 
 ### Patch Changes